seperate everything out
