+++ src/app/layout.tsx
@@ -1,14 +1,8 @@
 import type { Metadata } from "next";
-import { Space_Grotesk, Geo } from "next/font/google";
+import { Space_Grotesk } from "next/font/google";
 import "./globals.css";
 
 const spaceGrotesk = Space_Grotesk({
-  subsets: ["latin"],
-  display: "swap",
-});
-
-const geo = Geo({
-  weight: "400",
   subsets: ["latin"],
   display: "swap",
 });
