+++ src/utils/fonts.ts
@@ -1,0 +1,7 @@
+import { Geo } from "next/font/google";
+
+export const geo = Geo({
+  subsets: ["latin"],
+  weight: "400",
+  display: "swap",
+});
